+++ backend/canisters/group/api/src/updates/c2c_handle_bot_action.rs
@@ -1,0 +1,2 @@
+pub type Args = bot_api_gateway_canister::c2c_handle_bot_action::Args;
+pub type Response = bot_api_gateway_canister::c2c_handle_bot_action::Response;