+++ backend/canisters/group/api/src/updates/mod.rs
@@ -5,7 +5,7 @@ pub mod c2c_delete_group;
 pub mod c2c_export_group;
 pub mod c2c_export_group_events;
 pub mod c2c_freeze_group;
-pub mod c2c_handle_bot_actions;
+pub mod c2c_handle_bot_action;
 pub mod c2c_invite_users;
 pub mod c2c_join_group;
 pub mod c2c_leave_group;