+++ backend/libraries/constants/src/lib.rs
@@ -7,6 +7,7 @@ pub const HOUR_IN_MS: Milliseconds = MINUTE_IN_MS * 60;
 pub const DAY_IN_MS: Milliseconds = HOUR_IN_MS * 24;
 pub const WEEK_IN_MS: Milliseconds = DAY_IN_MS * 7;
 pub const NANOS_PER_MILLISECOND: u64 = 1_000_000;
+pub const ONE_MB: u32 = 1024 * 1024;
 
 // This only applies to the 'top level' canisters (ie. not user + group canisters)
 pub fn min_cycles_balance(test_mode: bool) -> Cycles {