+++ backend/libraries/group_chat_core/src/members.rs
@@ -4,9 +4,10 @@ use crate::mentions::Mentions;
 use crate::roles::GroupRoleInternal;
 use crate::AccessRulesInternal;
 use candid::Principal;
-use constants::calculate_summary_updates_data_removal_cutoff;
+use constants::{calculate_summary_updates_data_removal_cutoff, ONE_MB};
 use group_community_common::{Member, MemberUpdate, Members};
 use serde::{Deserialize, Serialize};
+use serde_bytes::ByteBuf;
 use std::cell::OnceCell;
 use std::cmp::max;
 use std::collections::{BTreeMap, BTreeSet, VecDeque};
@@ -136,6 +137,15 @@ impl GroupMembers {
 
     pub fn set_chat(&mut self, chat: MultiUserChat) {
         self.stable_memory_members_map.set_chat(chat);
+    }
+
+    pub fn read_members_as_bytes_from_stable_memory(&self, after: Option<UserId>) -> Vec<ByteBuf> {
+        self.stable_memory_members_map
+            .read_members_as_bytes(after, 2 * ONE_MB as usize)
+    }
+
+    pub fn write_members_from_bytes_to_stable_memory(chat: MultiUserChat, members: Vec<ByteBuf>) -> Option<UserId> {
+        stable_memory::write_members_from_bytes(chat, members)
     }
 
     pub fn add(